moving trunk development to stage
git-svn-id: https://stage.maemo.org/svn/maemo/projects/haf/trunk@1715 50eb543b-83fa-0310-9bb7-f7cbbee4ab39

+++ osso-app-killer/src/gconf-clean.sh
@@ -1,8 +1,13 @@
 #!/bin/sh
-for d in `ls /etc/osso-af-init/gconf-dir/`; do
-  if [ "x$d" = "xschemas" ]; then
-    continue
-  fi
-  echo "$0: removing GConf subdirectory $d"
-  rm -rf $d
-done
+cd /etc/osso-af-init/gconf-dir
+if [ $? = 0 ]; then
+  for d in `ls`; do
+    if [ "x$d" = "xschemas" ]; then
+      continue
+    fi
+    echo "$0: removing GConf subdirectory $d"
+    rm -rf $d
+  done
+else
+  echo "$0: 'cd' command failed, doing nothing"
+fi
